Fixups
- OSS stats do not encode the target name. Create a new struct `OssStat` instead of incorectly reusing `TargetStat`.
- OSS stats do not belong under ost_parser because they don't encode target specific information.
  Move ost_parser.rs to oss_parser.rs and move what was in oss_parser.rs into mod.rs.
- Cleanup parsers
- Add more const
- Delete unused snapshots.
