+++ src/brw_stats_parser.rs
@@ -7,7 +7,6 @@ use crate::{
     time::time_triple,
     types::{BrwStats, BrwStatsBucket},
 };
-
 use combine::{
     attempt, choice,
     error::ParseError,
